Extracted the classes `TokenPoolCache` and `TokenPoolCacheImpl` to their own files.
